Updated various files in accordance with newer projects (LCD, SD)

+++ STM32F4_Diode/src/stubs.c
@@ -66,7 +66,7 @@ int _write(int fileHandle, char *buf, int len) {
 
 	int i;
 	for (i=0; i<len; i++) {
-		USART2_Putc(buf[i]);
+		USART2_Putc((uint8_t)buf[i]);
 	}
 
 	return len;
